Fix(cornerstone): the circle adapter to prevent writing undefined (#266)
* fix(dcmjs):The area and diameter were undefined.

The area and diameter written to the DICOM SR had
the value "undefined" because neither item was being
computed.  Also, the diameter used the area as the
actual value (which happened to be undefined).  This
change reads the values from the source adapter and
then provides them to the TID.300 adapter in the correct
places.

* Removed an unnecessary console log

* fix(dcmjs):Add a comment about the 0 dummy values