feat: 🎸 add value multiplicity to PersonName tags
And also adds the concept of tag and value accessors to
ValueRepresentation. Whenever a tag is created, it is assigned to a
Proxy which allows values to be represented differently according to
whether it's being output to JSON or being interacted with. This allows
a user to assign a string to a PN tag, and have that tag be output as a
JSON object conformant to spec.