revert: Changes from streaming handler (#463)
There were issues found when running the streaming data
parser, and using the async handling in JavaScript seems to
produce an easier to understand version, so this change is being
reverted in favour of the easier to understand version.

A limited number of changes are kept for things like constants.